Add missing Unmap call
An Unmap call was missing after its matching Map call while copying
results back to the result buffer. The DX debug layer will consider this
to be an error. This can cause the args_f32 buffer to always contain
stale values on certain GPUs and is now fixed.

Tested by running a simplified mpm88 example on a discrete GPU.